get api_key from enviroenment

+++ downloadFromPipelineHistories.py
@@ -20,7 +20,7 @@ Given a list of regular expressions pulls the Nth (default is 24th) dataset from
                       help="URL of Galaxy API")
     parser.add_option('-k', '--api_key', default=None,
                       help="Galaxy API key for connecting. REQUIRED!")
-    parser.add_option('-d', '--dataset_index', default=24, type='int',
+    parser.add_option('-d', '--dataset_index', default=-1, type='int',
                       help="which dataset to get in each history")
     parser.add_option('-D', '--dataset_regex', default=None,
                       help="Expression for matching dataset names. If set, overrides the dataset_index (-d)")
@@ -40,15 +40,24 @@ Given a list of regular expressions pulls the Nth (default is 24th) dataset from
     setupLogging(options, description)
 
     if options.api_key is None:
-        parser.error("You must speicfy an API key with the -k flag!")
+        key = edl.galaxy.getApiKey()
+        if key is None:
+            parser.error("You must speicfy an API key with the -k flag!")
+        else:
+            options.api_key = key
 
     # set up dataset search values
     if options.dataset_regex is None:
+        if options.dataset_index == -1:
+            parser.error('Please supply a dataset number or dataset string with -d or -D!')
         dsNum=options.dataset_index
         dsName=None
     else:
         dsNum=None
         dsName=re.compile(options.dataset_regex)
+
+    logging.debug("Looking for dataset: %r/%r" % (options.dataset_regex, 
+                                                  options.dataset_index))
 
     # Are we on the same machine as galaxy?
     filesAreLocal = re.search(r'://localhost',options.api_url) is not None
@@ -66,7 +75,7 @@ Given a list of regular expressions pulls the Nth (default is 24th) dataset from
                 hdir, dataset)
 
         # Create symlink if we can find the file locally
-        if filesAreLocal and 'file_name' in dataset:
+        if filesAreLocal and 'file_name' in dataset and not options.force_copy:
             originalFile=dataset['file_name']
             os.symlink(originalFile, out_file_name)
         else:
